+++ lib/src/dependency.g.dart
@@ -1,6 +1,6 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
 
-// ignore_for_file: prefer_expression_function_bodies
+// ignore_for_file: deprecated_member_use_from_same_package, lines_longer_than_80_chars, prefer_expression_function_bodies
 
 part of 'dependency.dart';
 
@@ -10,12 +10,10 @@ part of 'dependency.dart';
 
 SdkDependency _$SdkDependencyFromJson(Map json) {
   return $checkedNew('SdkDependency', json, () {
-    $checkKeys(json,
-        requiredKeys: const ['sdk'], disallowNullValues: const ['sdk']);
     final val = SdkDependency(
       $checkedConvert(json, 'sdk', (v) => v as String),
       version: $checkedConvert(
-          json, 'version', (v) => _constraintFromString(v as String)),
+          json, 'version', (v) => _constraintFromString(v as String?)),
     );
     return val;
   });
@@ -23,12 +21,10 @@ SdkDependency _$SdkDependencyFromJson(Map json) {
 
 GitDependency _$GitDependencyFromJson(Map json) {
   return $checkedNew('GitDependency', json, () {
-    $checkKeys(json,
-        requiredKeys: const ['url'], disallowNullValues: const ['url']);
     final val = GitDependency(
       $checkedConvert(json, 'url', (v) => parseGitUri(v as String)),
-      $checkedConvert(json, 'ref', (v) => v as String),
-      $checkedConvert(json, 'path', (v) => v as String),
+      ref: $checkedConvert(json, 'ref', (v) => v as String?),
+      path: $checkedConvert(json, 'path', (v) => v as String?),
     );
     return val;
   });
@@ -41,9 +37,9 @@ HostedDependency _$HostedDependencyFromJson(Map json) {
         disallowNullValues: const ['hosted']);
     final val = HostedDependency(
       version: $checkedConvert(
-          json, 'version', (v) => _constraintFromString(v as String)),
-      hosted: $checkedConvert(
-          json, 'hosted', (v) => v == null ? null : HostedDetails.fromJson(v)),
+          json, 'version', (v) => _constraintFromString(v as String?)),
+      hosted: $checkedConvert(json, 'hosted',
+          (v) => v == null ? null : HostedDetails.fromJson(v as Object)),
     );
     return val;
   });
@@ -52,12 +48,10 @@ HostedDependency _$HostedDependencyFromJson(Map json) {
 HostedDetails _$HostedDetailsFromJson(Map json) {
   return $checkedNew('HostedDetails', json, () {
     $checkKeys(json,
-        allowedKeys: const ['name', 'url'],
-        requiredKeys: const ['name'],
-        disallowNullValues: const ['name', 'url']);
+        allowedKeys: const ['name', 'url'], disallowNullValues: const ['url']);
     final val = HostedDetails(
       $checkedConvert(json, 'name', (v) => v as String),
-      $checkedConvert(json, 'url', (v) => parseGitUri(v as String)),
+      $checkedConvert(json, 'url', (v) => parseGitUriOrNull(v as String?)),
     );
     return val;
   });