+++ lib/src/pubspec.dart
@@ -17,53 +17,57 @@ class Pubspec {
   final String name;
 
   @JsonKey(fromJson: _versionFromString)
-  final Version version;
+  final Version? version;
 
-  final String description;
+  final String? description;
 
   /// This should be a URL pointing to the website for the package.
-  final String homepage;
+  final String? homepage;
 
   /// Specifies where to publish this package.
   ///
   /// Accepted values: `null`, `'none'` or an `http` or `https` URL.
   ///
   /// [More information](https://dart.dev/tools/pub/pubspec#publish_to).
-  final String publishTo;
+  final String? publishTo;
 
   /// Optional field to specify the source code repository of the package.
   /// Useful when a package has both a home page and a repository.
-  final Uri repository;
+  final Uri? repository;
 
   /// Optional field to a web page where developers can report new issues or
   /// view existing ones.
-  final Uri issueTracker;
+  final Uri? issueTracker;
 
   /// If there is exactly 1 value in [authors], returns it.
   ///
   /// If there are 0 or more than 1, returns `null`.
   @Deprecated(
-      'Here for completeness, but not recommended. Use `authors` instead.')
-  String get author {
+    'See https://dart.dev/tools/pub/pubspec#authorauthors',
+  )
+  String? get author {
     if (authors.length == 1) {
       return authors.single;
     }
     return null;
   }
 
+  @Deprecated(
+    'See https://dart.dev/tools/pub/pubspec#authorauthors',
+  )
   final List<String> authors;
-  final String documentation;
+  final String? documentation;
 
   @JsonKey(fromJson: _environmentMap)
-  final Map<String, VersionConstraint> environment;
+  final Map<String, VersionConstraint?>? environment;
 
-  @JsonKey(fromJson: parseDeps, nullable: false)
+  @JsonKey(fromJson: parseDeps)
   final Map<String, Dependency> dependencies;
 
-  @JsonKey(fromJson: parseDeps, nullable: false)
+  @JsonKey(fromJson: parseDeps)
   final Map<String, Dependency> devDependencies;
 
-  @JsonKey(fromJson: parseDeps, nullable: false)
+  @JsonKey(fromJson: parseDeps)
   final Map<String, Dependency> dependencyOverrides;
 
   /// Optional configuration specific to [Flutter](https://flutter.io/)
@@ -72,7 +76,7 @@ class Pubspec {
   /// May include
   /// [assets](https://flutter.io/docs/development/ui/assets-and-images)
   /// and other settings.
-  final Map<String, dynamic> flutter;
+  final Map<String, dynamic>? flutter;
 
   /// If [author] and [authors] are both provided, their values are combined
   /// with duplicates eliminated.
@@ -80,30 +84,38 @@ class Pubspec {
     this.name, {
     this.version,
     this.publishTo,
-    String author,
-    List<String> authors,
-    Map<String, VersionConstraint> environment,
+    @Deprecated(
+      'See https://dart.dev/tools/pub/pubspec#authorauthors',
+    )
+        String? author,
+    @Deprecated(
+      'See https://dart.dev/tools/pub/pubspec#authorauthors',
+    )
+        List<String>? authors,
+    Map<String, VersionConstraint?>? environment,
     this.homepage,
     this.repository,
     this.issueTracker,
     this.documentation,
     this.description,
-    Map<String, Dependency> dependencies,
-    Map<String, Dependency> devDependencies,
-    Map<String, Dependency> dependencyOverrides,
+    Map<String, Dependency>? dependencies,
+    Map<String, Dependency>? devDependencies,
+    Map<String, Dependency>? dependencyOverrides,
     this.flutter,
-  })  : authors = _normalizeAuthors(author, authors),
+  })  :
+        // ignore: deprecated_member_use_from_same_package
+        authors = _normalizeAuthors(author, authors),
         environment = environment ?? const {},
         dependencies = dependencies ?? const {},
         devDependencies = devDependencies ?? const {},
         dependencyOverrides = dependencyOverrides ?? const {} {
-    if (name == null || name.isEmpty) {
+    if (name.isEmpty) {
       throw ArgumentError.value(name, 'name', '"name" cannot be empty.');
     }
 
     if (publishTo != null && publishTo != 'none') {
       try {
-        final targetUri = Uri.parse(publishTo);
+        final targetUri = Uri.parse(publishTo!);
         if (!(targetUri.isScheme('http') || targetUri.isScheme('https'))) {
           throw const FormatException('Must be an http or https URL.');
         }
@@ -114,8 +126,6 @@ class Pubspec {
   }
 
   factory Pubspec.fromJson(Map json, {bool lenient = false}) {
-    lenient ??= false;
-
     if (lenient) {
       while (json.isNotEmpty) {
         // Attempting to remove top-level properties that cause parsing errors.
@@ -138,30 +148,26 @@ class Pubspec {
   ///
   /// When [lenient] is set, top-level property-parsing or type cast errors are
   /// ignored and `null` values are returned.
-  factory Pubspec.parse(String yaml, {sourceUrl, bool lenient = false}) {
-    lenient ??= false;
+  factory Pubspec.parse(String yaml, {Uri? sourceUrl, bool lenient = false}) =>
+      checkedYamlDecode(
+        yaml,
+        (map) => Pubspec.fromJson(map!, lenient: lenient),
+        sourceUrl: sourceUrl,
+      );
 
-    return checkedYamlDecode(
-        yaml, (map) => Pubspec.fromJson(map, lenient: lenient),
-        sourceUrl: sourceUrl);
-  }
-
-  static List<String> _normalizeAuthors(String author, List<String> authors) {
-    final value = <String>{};
-    if (author != null) {
-      value.add(author);
-    }
-    if (authors != null) {
-      value.addAll(authors);
-    }
+  static List<String> _normalizeAuthors(String? author, List<String>? authors) {
+    final value = <String>{
+      if (author != null) author,
+      ...?authors,
+    };
     return value.toList();
   }
 }
 
-Version _versionFromString(String input) =>
+Version? _versionFromString(String? input) =>
     input == null ? null : Version.parse(input);
 
-Map<String, VersionConstraint> _environmentMap(Map source) =>
+Map<String, VersionConstraint?>? _environmentMap(Map? source) =>
     source?.map((k, value) {
       final key = k as String;
       if (key == 'dart') {
@@ -176,7 +182,7 @@ Map<String, VersionConstraint> _environmentMap(Map source) =>
         );
       }
 
-      VersionConstraint constraint;
+      VersionConstraint? constraint;
       if (value == null) {
         constraint = null;
       } else if (value is String) {